Echo/backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
deepgram-sdk
google-genai
httpx
